Rename app directories: use snake case
