+++ geometry_game/point.py
