+++ geometry_game/rectangle.py
